refactor: 중복을 검사하는 로직을 Service 계층으로 통일
