+++ src/main/java/cart/dao/JdbcCartDao.java
@@ -2,8 +2,6 @@ package cart.dao;
 
 import cart.entity.Cart;
 import cart.entity.PutCart;
-import cart.exception.DaoDuplicateException;
-import cart.exception.ServiceIllegalArgumentException;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
@@ -13,8 +11,6 @@ import java.util.List;
 @Repository
 public class JdbcCartDao implements CartDao {
 
-    private static final String EXISTS_ITEM_IN_CART_MESSAGE = "이미 장바구니에 담은 상품입니다.";
-
     private final JdbcTemplate jdbcTemplate;
 
     public JdbcCartDao(JdbcTemplate jdbcTemplate) {
@@ -23,9 +19,6 @@ public class JdbcCartDao implements CartDao {
 
     @Override
     public void save(PutCart putCart) {
-        if (isCartExists(putCart)) {
-            throw new DaoDuplicateException(EXISTS_ITEM_IN_CART_MESSAGE);
-        }
         String sql = "insert into cart(member_id, item_id) values(?, ?)";
 
         jdbcTemplate.update(sql, putCart.getMemberId(), putCart.getProductId());
