--- src/main/java/cart/exception/DaoDuplicateException.java
@@ -1,8 +1,0 @@
-package cart.exception;
-
-public class DaoDuplicateException extends RuntimeException{
-
-    public DaoDuplicateException(String message) {
-        super(message);
-    }
-}
